Query actions constants moved to seperate file in constants folder
