+++ app/constants/query.js
@@ -1,0 +1,11 @@
+export const QUERY_SET = 'QUERY_SET'
+export const QUERY_SET_RESULTS = 'QUERY_SET_RESULTS'
+export const QUERY_RUN_REQUEST = 'QUERY_RUN_REQUEST'
+export const QUERY_RUN_SUCCESS = 'QUERY_RUN_SUCCESS'
+export const QUERY_RUN_FAILURE = 'QUERY_RUN_FAILURE'
+export const QUERY_REQUEST = 'QUERY_REQUEST'
+export const QUERY_SUCCESS = 'QUERY_SUCCESS'
+export const QUERY_FAILURE = 'QUERY_FAILURE'
+export const QUERIES_REQUEST = 'QUERIES_REQUEST'
+export const QUERIES_SUCCESS = 'QUERIES_SUCCESS'
+export const QUERIES_FAILURE = 'QUERIES_FAILURE'
